Unify SecuritySystemCurrentState and SecuritySystemTargetState into SecuritySystemState
